feat: update example app to use SvelteKit remote functions
- Update SvelteKit to version 2.27.0 to support remote functions

- Enable experimental remote functions and async await in svelte.config.js

- Replace traditional loader pattern with remote function in examples

- Create tweets.remote.ts to fetch example tweets using query function

- Update +page.svelte to use await syntax for cleaner code
