+++ package.json
@@ -52,7 +52,7 @@
 		"preinstall": "npx only-allow pnpm"
 	},
 	"peerDependencies": {
-		"@sveltejs/kit": "^2.26.1",
+		"@sveltejs/kit": "^2.27.0",
 		"svelte": "^5.36.17"
 	},
 	"devDependencies": {
@@ -61,7 +61,7 @@
 		"@ryoppippi/svelte-preprocess-import-css": "jsr:^0.3.4",
 		"@std/async": "jsr:^1.0.12",
 		"@sveltejs/adapter-vercel": "^5.8.1",
-		"@sveltejs/kit": "^2.26.1",
+		"@sveltejs/kit": "^2.27.0",
 		"@sveltejs/package": "^2.4.0",
 		"@sveltejs/vite-plugin-svelte": "5.1.1",
 		"bumpp": "^10.2.0",
