+++ svelte.config.js
@@ -32,6 +32,9 @@ const config = {
 
 	compilerOptions: {
 		runes: true,
+		experimental: {
+			async: true,
+		},
 	},
 
 	vitePlugin: {
@@ -43,6 +46,10 @@ const config = {
 		// If your environment is not supported or you settled on a specific environment, switch out the adapter.
 		// See https://kit.svelte.dev/docs/adapters for more information about adapters.
 		adapter: adapter(),
+
+		experimental: {
+			remoteFunctions: true,
+		},
 
 		alias: {
 			$rt: relativePath('./src/lib/react-tweet'),
